refactor(client): move menu logic out of adminpage
