+++ packages/client/app/Router.tsx
@@ -1,5 +1,5 @@
-import { type ReactNode } from 'react'
-import { Route, Routes, Navigate } from 'react-router-dom'
+import { type ReactNode, useContext } from 'react'
+import { Route, Routes, Navigate, Outlet } from 'react-router-dom'
 
 import AcceptArticleOwnershipPage from './components/component-dashboard/src/components/AcceptArticleOwnershipPage'
 import ArticleArtifactPage from './components/component-published-artifact/components/ArticleArtifactPage'
@@ -7,12 +7,17 @@ import DeclineArticleOwnershipPage from './components/component-dashboard/src/co
 import GroupPage from './components/component-frontpage/src/GroupPage'
 import InvitationAcceptedPage from './components/component-dashboard/src/components/InvitationAcceptedPage'
 import Login from './components/component-login/src'
+import { XpubContext } from './components/xpub-with-context/src'
 
-import { GroupWrapperPage } from './pages'
+import { GroupWrapperPage, MenuPage } from './pages'
+
+import Layout from './ui/base/Layout'
 
 import AdminPage from './components/AdminPage'
 
 const Router = (): ReactNode => {
+  const [conversion] = useContext(XpubContext)
+
   return (
     <Routes>
       <Route element={<GroupPage />} path="/" />
@@ -42,8 +47,16 @@ const Router = (): ReactNode => {
           path={`versions/:version/artifacts/:artifactId`}
         />
 
-        {/* AdminPage has nested routes within */}
-        <Route element={<AdminPage />} path="*" />
+        <Route
+          element={
+            <Layout converting={conversion.converting}>
+              <MenuPage />
+              <Outlet />
+            </Layout>
+          }
+        >
+          <Route element={<AdminPage />} path="*" />
+        </Route>
       </Route>
     </Routes>
   )
