+++ packages/client/app/components/AdminPage.jsx
@@ -1,23 +1,14 @@
 /* eslint-disable react/prop-types, react-hooks/exhaustive-deps, react-hooks/immutability, react-hooks/refs */
 
-import { useApolloClient, useMutation, useQuery } from '@apollo/client/react'
+import { useMutation, useQuery } from '@apollo/client/react'
 import PropTypes from 'prop-types'
 import { useContext, useEffect, useRef } from 'react'
 import Modal from 'react-modal'
-import {
-  matchPath,
-  Navigate,
-  Route,
-  Routes,
-  useLocation,
-} from 'react-router-dom'
+import { Navigate, Route, Routes } from 'react-router-dom'
 import i18next from 'i18next'
-import { useTranslation } from 'react-i18next'
 import { JournalContext } from './xpub-journal'
-import { XpubContext } from './xpub-with-context/src'
 import { ConfigContext } from './config/src'
 import { getLanguages } from '../i18n'
-import Layout from '../ui/base/Layout'
 
 import FormBuilderPage from './component-formbuilder/src/components/FormBuilderPage'
 import ManuscriptPage from './component-manuscript/src/components/ManuscriptPage'
@@ -42,17 +33,12 @@ import CMSPublishingCollectionPage from './component-cms-manager/src/CMSPublishi
 import CMSFileBrowserPage from './component-cms-manager/src/CMSFileBrowserPage'
 import { CURRENT_USER, UPDATE_LANGUAGE } from '../queries'
 
-import Menu from './Menu'
 import { Spinner, PageError } from './shared'
 
 import DashboardEditsPage from './component-dashboard/src/components/DashboardEditsPage'
 import DashboardLayout from './component-dashboard/src/components/DashboardLayout'
 import DashboardReviewsPage from './component-dashboard/src/components/DashboardReviewsPage'
 import DashboardSubmissionsPage from './component-dashboard/src/components/DashboardSubmissionsPage'
-
-const getParams = ({ routerPath, path }) => {
-  return matchPath(routerPath, path).params
-}
 
 // TODO: Redirect if token expires
 const PrivateRoute = ({
@@ -89,14 +75,10 @@ PrivateRoute.propTypes = {
 const AdminPage = () => {
   Modal.setAppElement('#root')
   const journal = useContext(JournalContext)
-  const [conversion] = useContext(XpubContext)
   const config = useContext(ConfigContext)
-  const { urlFrag, instanceName } = config
-  const location = useLocation()
-  const { t } = useTranslation()
-  const client = useApolloClient()
+  const { urlFrag } = config
 
-  const { loading, error, data } = useQuery(CURRENT_USER, {
+  const { loading, data } = useQuery(CURRENT_USER, {
     fetchPolicy: 'network-only',
     pollInterval: 120000,
   })
@@ -126,185 +108,29 @@ const AdminPage = () => {
     return <Spinner />
   }
 
-  let notice = ''
-
-  if (error || !data?.currentUser) {
-    if (error?.networkError) {
-      notice = 'You are offline.'
-    } else {
-      if (localStorage.getItem('token') !== null) {
-        localStorage.removeItem('token')
-      }
-
-      client.cache.reset()
-
-      localStorage.setItem('intendedPage', location.pathname + location.search)
-      const redirectlocation = `${urlFrag}/login`
-      return <Navigate replace to={redirectlocation} />
-    }
-  }
-
   const currentUser = data?.currentUser
   journal.textStyles = data?.builtCss?.css
-  const hasAlert = data?.userHasTaskAlerts
 
   previousDataRef.current = data
 
-  const showLinks = location.pathname.match(/^\/(submit|manuscript)/g)
-  let links = []
-  const submissionFormBuilderLink = `${urlFrag}/admin/submission-form-builder`
-  const reviewFormBuilderLink = `${urlFrag}/admin/review-form-builder`
-  const decisionFormBuilderLink = `${urlFrag}/admin/decision-form-builder`
-  const configurationLink = `${urlFrag}/admin/configuration`
   const homeLink = `${urlFrag}/dashboard`
   const dashboardSubmissionsLink = `${urlFrag}/dashboard/submissions`
   const dashboardReviewsLink = `${urlFrag}/dashboard/reviews`
   const dashboardEditsLink = `${urlFrag}/dashboard/edits`
-  const profileLink = `${urlFrag}/profile`
-  const manuscriptsLink = `${urlFrag}/admin/manuscripts`
   const reportsLink = `${urlFrag}/admin/reports`
-  const userAdminLink = `${urlFrag}/admin/users`
-  const tasksTemplateLink = `${urlFrag}/admin/tasks`
+  const profileLink = `${urlFrag}/profile`
   const CMSPagesPageLink = `${urlFrag}/admin/cms/pages`
   const CMSLayoutPageLink = `${urlFrag}/admin/cms/layout`
   const CMSArticlePageLink = `${urlFrag}/admin/cms/article`
   const CMSFileBrowserLink = `${urlFrag}/admin/cms/filebrowser`
   const CMSMetadataPageLink = `${urlFrag}/admin/cms/metadata`
   const CMSPublishingCollectionPageLink = `${urlFrag}/admin/cms/collections`
-  const loginLink = `${urlFrag}/login?next=${homeLink}`
-  const path = `${urlFrag}/versions/:version`
   const redirectLink = `${urlFrag}/login?next=${homeLink}`
-
-  if (showLinks) {
-    const params = getParams(location.pathname, path)
-    const baseLink = `${urlFrag}/versions/${params.version}`
-    const submitLink = `${baseLink}/submit`
-    const manuscriptLink = `${baseLink}/manuscript`
-
-    links = showLinks
-      ? [
-          { link: submitLink, name: t('leftMenu.Summary Info') },
-          {
-            link: manuscriptLink,
-            name: t('leftMenu.Manuscript'),
-          },
-        ]
-      : null
-  }
 
   const isUser = currentUser?.groupRoles?.includes('user')
   const isGroupManager = currentUser?.groupRoles?.includes('groupManager')
   const isGroupAdmin = currentUser?.groupRoles?.includes('groupAdmin')
   const isAdmin = currentUser?.globalRoles?.includes('admin')
-
-  if (
-    currentUser &&
-    (isUser || isGroupManager || isGroupAdmin || isAdmin) &&
-    ['journal', 'prc', 'preprint2'].includes(instanceName) // TODO: remove instance based logic and refactor it to be enabled and disabled from config manager
-  ) {
-    links.push({
-      link: homeLink,
-      name: t('leftMenu.Dashboard'),
-      icon: 'home',
-      hasAlert,
-    })
-  }
-
-  if (isGroupManager || isGroupAdmin) {
-    links.push({
-      link: manuscriptsLink,
-      name: t('leftMenu.Manuscripts'),
-      icon: 'file-text',
-    })
-    if (config?.report?.showInMenu && isGroupAdmin)
-      links.push({
-        link: reportsLink,
-        name: t('leftMenu.Reports'),
-        icon: 'activity',
-      })
-  }
-
-  if (isGroupAdmin || isAdmin) {
-    links.push({
-      menu: 'Settings',
-      name: t('leftMenu.Settings'),
-      icon: 'settings',
-      links: [
-        {
-          menu: 'Forms',
-          name: t('leftMenu.Forms'),
-          icon: 'check-square',
-          links: [
-            {
-              link: submissionFormBuilderLink,
-              name: t('leftMenu.Submission'),
-            },
-            {
-              link: reviewFormBuilderLink,
-              name: t('leftMenu.Review'),
-            },
-            {
-              link: decisionFormBuilderLink,
-              name: t('leftMenu.Decision'),
-            },
-          ],
-        },
-        {
-          link: tasksTemplateLink,
-          name: t('leftMenu.Tasks'),
-          icon: 'list',
-        },
-        {
-          link: userAdminLink,
-          name: t('leftMenu.Users'),
-          icon: 'users',
-        },
-        {
-          link: configurationLink,
-          name: t('leftMenu.Configuration'),
-          icon: 'sliders',
-        },
-        {
-          menu: 'CMS',
-          name: t('leftMenu.CMS'),
-          icon: 'layout',
-          links: [
-            {
-              link: CMSPagesPageLink,
-              name: t('leftMenu.Pages'),
-              icon: '',
-            },
-            {
-              link: CMSLayoutPageLink,
-              name: t('leftMenu.Layout'),
-              icon: '',
-            },
-            {
-              link: CMSArticlePageLink,
-              name: t('leftMenu.Article'),
-              icon: '',
-            },
-            {
-              link: CMSMetadataPageLink,
-              name: t('leftMenu.Metadata'),
-              icon: '',
-            },
-            {
-              link: CMSFileBrowserLink,
-              name: t('leftMenu.FileBrowser'),
-
-              icon: '',
-            },
-            {
-              link: CMSPublishingCollectionPageLink,
-              name: t('leftMenu.Collections'),
-              icon: '',
-            },
-          ],
-        },
-      ],
-    })
-  }
 
   const invitationId = window.localStorage.getItem('invitationId')
     ? window.localStorage.getItem('invitationId')
@@ -333,395 +159,383 @@ const AdminPage = () => {
     )
 
   return (
-    <Layout $converting={conversion.converting}>
-      <Menu
-        brand={config?.groupIdentity?.brandName}
-        brandLink={homeLink}
-        className=""
-        loginLink={loginLink}
-        navLinkComponents={links}
-        notice={notice}
-        profileLink={profileLink}
-        user={currentUser}
+    <Routes>
+      <Route
+        element={
+          <PrivateRoute
+            component={ProfilePage}
+            currentUser={currentUser}
+            path={profileLink}
+            redirectLink={redirectLink}
+          />
+        }
+        path="profile"
       />
-      <Routes>
+
+      {(isUser || isGroupManager || isGroupAdmin || isAdmin) && [
         <Route
           element={
             <PrivateRoute
               component={ProfilePage}
               currentUser={currentUser}
-              path={profileLink}
+              path={`${urlFrag}/profile/:id`}
               redirectLink={redirectLink}
             />
           }
-          path="profile"
-        />
-
-        {(isUser || isGroupManager || isGroupAdmin || isAdmin) && [
-          <Route
-            element={
-              <PrivateRoute
-                component={ProfilePage}
-                currentUser={currentUser}
-                path={`${urlFrag}/profile/:id`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="profile"
-            path="profile/:id"
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                component={NewSubmissionPage}
-                currentUser={currentUser}
-                path={`${urlFrag}/newSubmission`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="new-submission"
-            path="newSubmission"
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                component={ReviewPage}
-                currentUser={currentUser}
-                path={`${urlFrag}/versions/:version/review`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="review"
-            path="versions/:version/review"
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                component={ReviewPreviewPage}
-                currentUser={currentUser}
-                path={`${urlFrag}/versions/:version/reviewPreview`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="review-preview"
-            path="versions/:version/reviewPreview"
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                component={DecisionPage}
-                currentUser={currentUser}
-                path={`${urlFrag}/versions/:version/decision`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="decision"
-            path="versions/:version/decision"
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                component={SubmitPage}
-                currentUser={currentUser}
-                path={`${urlFrag}/versions/:version/${
-                  ['preprint1', 'preprint2'].includes(config.instanceName)
-                    ? 'evaluation'
-                    : 'submit'
-                }`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="submit"
-            path={`versions/:version/${
-              ['preprint1', 'preprint2'].includes(config.instanceName)
-                ? 'evaluation'
-                : 'submit'
-            }`} // TODO: Remove instance based custom submit page and refactor it use config manager flag in future
-          />,
-          <Route
-            element={
-              <DashboardLayout urlFrag={urlFrag}>
-                <Routes>
+          key="profile"
+          path="profile/:id"
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              component={NewSubmissionPage}
+              currentUser={currentUser}
+              path={`${urlFrag}/newSubmission`}
+              redirectLink={redirectLink}
+            />
+          }
+          key="new-submission"
+          path="newSubmission"
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              component={ReviewPage}
+              currentUser={currentUser}
+              path={`${urlFrag}/versions/:version/review`}
+              redirectLink={redirectLink}
+            />
+          }
+          key="review"
+          path="versions/:version/review"
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              component={ReviewPreviewPage}
+              currentUser={currentUser}
+              path={`${urlFrag}/versions/:version/reviewPreview`}
+              redirectLink={redirectLink}
+            />
+          }
+          key="review-preview"
+          path="versions/:version/reviewPreview"
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              component={DecisionPage}
+              currentUser={currentUser}
+              path={`${urlFrag}/versions/:version/decision`}
+              redirectLink={redirectLink}
+            />
+          }
+          key="decision"
+          path="versions/:version/decision"
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              component={SubmitPage}
+              currentUser={currentUser}
+              path={`${urlFrag}/versions/:version/${
+                ['preprint1', 'preprint2'].includes(config.instanceName)
+                  ? 'evaluation'
+                  : 'submit'
+              }`}
+              redirectLink={redirectLink}
+            />
+          }
+          key="submit"
+          path={`versions/:version/${
+            ['preprint1', 'preprint2'].includes(config.instanceName)
+              ? 'evaluation'
+              : 'submit'
+          }`} // TODO: Remove instance based custom submit page and refactor it use config manager flag in future
+        />,
+        <Route
+          element={
+            <DashboardLayout urlFrag={urlFrag}>
+              <Routes>
+                <Route
+                  element={
+                    <PrivateRoute
+                      component={dashboardRedirect}
+                      currentUser={currentUser}
+                      path={homeLink}
+                      redirectLink={redirectLink}
+                    />
+                  }
+                  path=""
+                />
+                {config?.dashboard?.showSections?.includes('submission') && (
                   <Route
                     element={
                       <PrivateRoute
-                        component={dashboardRedirect}
+                        component={DashboardSubmissionsPage}
                         currentUser={currentUser}
-                        path={homeLink}
+                        path={dashboardSubmissionsLink}
                         redirectLink={redirectLink}
                       />
                     }
-                    path=""
+                    key="submission"
+                    path="submissions"
                   />
-                  {config?.dashboard?.showSections?.includes('submission') && (
-                    <Route
-                      element={
-                        <PrivateRoute
-                          component={DashboardSubmissionsPage}
-                          currentUser={currentUser}
-                          path={dashboardSubmissionsLink}
-                          redirectLink={redirectLink}
-                        />
-                      }
-                      key="submission"
-                      path="submissions"
-                    />
-                  )}
-                  {config?.dashboard?.showSections?.includes('review') && (
-                    <Route
-                      element={
-                        <PrivateRoute
-                          component={DashboardReviewsPage}
-                          currentUser={currentUser}
-                          path={dashboardReviewsLink}
-                          redirectLink={redirectLink}
-                        />
-                      }
-                      key="review"
-                      path="reviews"
-                    />
-                  )}
-                  {config?.dashboard?.showSections?.includes('editor') && (
-                    <Route
-                      element={
-                        <PrivateRoute
-                          component={DashboardEditsPage}
-                          currentUser={currentUser}
-                          path={dashboardEditsLink}
-                          redirectLink={redirectLink}
-                        />
-                      }
-                      key="editor"
-                      path="edits"
-                    />
-                  )}
-                </Routes>
-              </DashboardLayout>
-            }
-            key="dashboard"
-            path={`dashboard/*`}
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                component={ProductionPage}
-                currentUser={currentUser}
-                path={`${urlFrag}/versions/:version/production`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="production"
-            path="versions/:version/production"
-          />,
-        ]}
-        {(isGroupAdmin || isAdmin) && [
-          // We use array instead of <></> because of https://stackoverflow.com/a/68637108/6505513
-          <Route
-            element={
-              <PrivateRoute
-                component={FormBuilderPage}
-                currentUser={currentUser}
-                path={`${urlFrag}/admin/form-builder`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="form-builder"
-            path={`/admin/form-builder`}
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                category="submission"
-                component={FormBuilderPage}
-                currentUser={currentUser}
-                path={`${urlFrag}/admin/submission-form-builder`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="submission-form-builder"
-            path={`admin/submission-form-builder`}
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                category="review"
-                component={FormBuilderPage}
-                currentUser={currentUser}
-                path={`${urlFrag}/admin/review-form-builder`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="review-form-builder"
-            path={`admin/review-form-builder`}
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                category="decision"
-                component={FormBuilderPage}
-                currentUser={currentUser}
-                path={`${urlFrag}/admin/decision-form-builder`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="decision-form-builder"
-            path={`admin/decision-form-builder`}
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                component={UsersPage}
-                currentUser={currentUser}
-                path={`${urlFrag}/admin/users`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="users"
-            path={`admin/users`}
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                component={TasksTemplatePage}
-                path={`${urlFrag}/admin/tasks`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="tasks"
-            path={`admin/tasks`}
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                component={CMSPagesPage}
-                currentUser={currentUser}
-                path={`${CMSPagesPageLink}/:pageId?`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="CMSPagesPage"
-            path={`admin/cms/pages/:pageId?`}
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                component={CMSLayoutPage}
-                currentUser={currentUser}
-                path={CMSLayoutPageLink}
-                redirectLink={redirectLink}
-              />
-            }
-            key="CMSLayoutPage"
-            path="admin/cms/layout"
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                component={CMSArticlePage}
-                currentUser={currentUser}
-                path={CMSArticlePageLink}
-                redirectLink={redirectLink}
-              />
-            }
-            key="CMSArticlePage"
-            path="admin/cms/article"
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                component={CMSFileBrowserPage}
-                currentUser={currentUser}
-                path={CMSFileBrowserLink}
-                redirectLink={redirectLink}
-              />
-            }
-            key="CMSFileBrowserPage"
-            path="admin/cms/filebrowser"
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                component={CMSMetadataPage}
-                currentUser={currentUser}
-                path={CMSMetadataPageLink}
-                redirectLink={redirectLink}
-              />
-            }
-            key="CMSMetadataPage"
-            path="admin/cms/metadata"
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                component={CMSPublishingCollectionPage}
-                currentUser={currentUser}
-                path={CMSPublishingCollectionPageLink}
-                redirectLink={redirectLink}
-              />
-            }
-            key="CMSPublishingCollectionPage"
-            path="admin/cms/collections"
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                component={ConfigManagerPage}
-                path={`${urlFrag}/admin/configuration`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="configuration"
-            path={`admin/configuration`}
-          />,
-        ]}
-        {(isGroupManager || isGroupAdmin) && [
-          <Route
-            element={
-              <PrivateRoute
-                component={ManuscriptPage}
-                currentUser={currentUser}
-                path={`${urlFrag}/versions/:version/manuscript`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="manuscript"
-            path={`versions/:version/manuscript`}
-          />,
-          <Route
-            element={
-              <PrivateRoute
-                component={ManuscriptsPage}
-                currentUser={currentUser}
-                path={`${urlFrag}/admin/manuscripts`}
-                redirectLink={redirectLink}
-              />
-            }
-            key="manuscripts"
-            path={`admin/manuscripts`}
-          />,
-        ]}
-        {isGroupAdmin && [
-          <Route
-            element={
-              <PrivateRoute
-                component={ReportPage}
-                currentUser={currentUser}
-                path={reportsLink}
-                redirectLink={redirectLink}
-              />
-            }
-            key="reports"
-            path="admin/reports"
-          />,
-        ]}
+                )}
+                {config?.dashboard?.showSections?.includes('review') && (
+                  <Route
+                    element={
+                      <PrivateRoute
+                        component={DashboardReviewsPage}
+                        currentUser={currentUser}
+                        path={dashboardReviewsLink}
+                        redirectLink={redirectLink}
+                      />
+                    }
+                    key="review"
+                    path="reviews"
+                  />
+                )}
+                {config?.dashboard?.showSections?.includes('editor') && (
+                  <Route
+                    element={
+                      <PrivateRoute
+                        component={DashboardEditsPage}
+                        currentUser={currentUser}
+                        path={dashboardEditsLink}
+                        redirectLink={redirectLink}
+                      />
+                    }
+                    key="editor"
+                    path="edits"
+                  />
+                )}
+              </Routes>
+            </DashboardLayout>
+          }
+          key="dashboard"
+          path={`dashboard/*`}
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              component={ProductionPage}
+              currentUser={currentUser}
+              path={`${urlFrag}/versions/:version/production`}
+              redirectLink={redirectLink}
+            />
+          }
+          key="production"
+          path="versions/:version/production"
+        />,
+      ]}
+      {(isGroupAdmin || isAdmin) && [
+        // We use array instead of <></> because of https://stackoverflow.com/a/68637108/6505513
+        <Route
+          element={
+            <PrivateRoute
+              component={FormBuilderPage}
+              currentUser={currentUser}
+              path={`${urlFrag}/admin/form-builder`}
+              redirectLink={redirectLink}
+            />
+          }
+          key="form-builder"
+          path={`/admin/form-builder`}
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              category="submission"
+              component={FormBuilderPage}
+              currentUser={currentUser}
+              path={`${urlFrag}/admin/submission-form-builder`}
+              redirectLink={redirectLink}
+            />
+          }
+          key="submission-form-builder"
+          path={`admin/submission-form-builder`}
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              category="review"
+              component={FormBuilderPage}
+              currentUser={currentUser}
+              path={`${urlFrag}/admin/review-form-builder`}
+              redirectLink={redirectLink}
+            />
+          }
+          key="review-form-builder"
+          path={`admin/review-form-builder`}
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              category="decision"
+              component={FormBuilderPage}
+              currentUser={currentUser}
+              path={`${urlFrag}/admin/decision-form-builder`}
+              redirectLink={redirectLink}
+            />
+          }
+          key="decision-form-builder"
+          path={`admin/decision-form-builder`}
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              component={UsersPage}
+              currentUser={currentUser}
+              path={`${urlFrag}/admin/users`}
+              redirectLink={redirectLink}
+            />
+          }
+          key="users"
+          path={`admin/users`}
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              component={TasksTemplatePage}
+              path={`${urlFrag}/admin/tasks`}
+              redirectLink={redirectLink}
+            />
+          }
+          key="tasks"
+          path={`admin/tasks`}
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              component={CMSPagesPage}
+              currentUser={currentUser}
+              path={`${CMSPagesPageLink}/:pageId?`}
+              redirectLink={redirectLink}
+            />
+          }
+          key="CMSPagesPage"
+          path={`admin/cms/pages/:pageId?`}
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              component={CMSLayoutPage}
+              currentUser={currentUser}
+              path={CMSLayoutPageLink}
+              redirectLink={redirectLink}
+            />
+          }
+          key="CMSLayoutPage"
+          path="admin/cms/layout"
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              component={CMSArticlePage}
+              currentUser={currentUser}
+              path={CMSArticlePageLink}
+              redirectLink={redirectLink}
+            />
+          }
+          key="CMSArticlePage"
+          path="admin/cms/article"
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              component={CMSFileBrowserPage}
+              currentUser={currentUser}
+              path={CMSFileBrowserLink}
+              redirectLink={redirectLink}
+            />
+          }
+          key="CMSFileBrowserPage"
+          path="admin/cms/filebrowser"
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              component={CMSMetadataPage}
+              currentUser={currentUser}
+              path={CMSMetadataPageLink}
+              redirectLink={redirectLink}
+            />
+          }
+          key="CMSMetadataPage"
+          path="admin/cms/metadata"
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              component={CMSPublishingCollectionPage}
+              currentUser={currentUser}
+              path={CMSPublishingCollectionPageLink}
+              redirectLink={redirectLink}
+            />
+          }
+          key="CMSPublishingCollectionPage"
+          path="admin/cms/collections"
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              component={ConfigManagerPage}
+              path={`${urlFrag}/admin/configuration`}
+              redirectLink={redirectLink}
+            />
+          }
+          key="configuration"
+          path={`admin/configuration`}
+        />,
+      ]}
+      {(isGroupManager || isGroupAdmin) && [
+        <Route
+          element={
+            <PrivateRoute
+              component={ManuscriptPage}
+              currentUser={currentUser}
+              path={`${urlFrag}/versions/:version/manuscript`}
+              redirectLink={redirectLink}
+            />
+          }
+          key="manuscript"
+          path={`versions/:version/manuscript`}
+        />,
+        <Route
+          element={
+            <PrivateRoute
+              component={ManuscriptsPage}
+              currentUser={currentUser}
+              path={`${urlFrag}/admin/manuscripts`}
+              redirectLink={redirectLink}
+            />
+          }
+          key="manuscripts"
+          path={`admin/manuscripts`}
+        />,
+      ]}
+      {isGroupAdmin && [
+        <Route
+          element={
+            <PrivateRoute
+              component={ReportPage}
+              currentUser={currentUser}
+              path={reportsLink}
+              redirectLink={redirectLink}
+            />
+          }
+          key="reports"
+          path="admin/reports"
+        />,
+      ]}
 
-        {isUser || isGroupManager || isGroupAdmin || isAdmin ? (
-          <Route element={<PageError errorCode={404} />} path="*" />
-        ) : (
-          <Route element={<PageError errorCode={403} />} path="*" />
-        )}
-      </Routes>
-    </Layout>
+      {isUser || isGroupManager || isGroupAdmin || isAdmin ? (
+        <Route element={<PageError errorCode={404} />} path="*" />
+      ) : (
+        <Route element={<PageError errorCode={403} />} path="*" />
+      )}
+    </Routes>
   )
 }
