+++ packages/client/app/pages/index.ts
@@ -1,1 +1,2 @@
 export { default as GroupWrapperPage } from './GroupWrapper.page'
+export { default as MenuPage } from './Menu.page'
